{"types":"C","runs":19}

+++ L7Sl7.java
@@ -11,7 +11,7 @@ public class L7Sl7 extends GraphicsProgram {
 	public void run() {
 		for (int i = 0; i < N_ROWS; i++) {
 			for (int j = 0; j < N_COLUMNS; j++) {
-				double squareWidth = getHeight() / N_ROWS;
+				double squareWidth = (double) getHeight() / N_ROWS;
 				GRect square = new GRect(squareWidth, squareWidth);
 				double x = squareWidth * j;
 				double y = squareWidth * i;
